add searchByDataMasking

+++ src/main/java/cn/com/taiji/elasticsearchclient/dto/CorporationDTO.java
@@ -20,6 +20,6 @@ public class CorporationDTO implements Serializable {
 
     private Long totalNum;
 
-    private double totalPage;
+    private Integer totalPage;
 
 }
